refactor(ui): toggle component
